requirements/basic-requirements.txt
asgiref==3.6.0
Django==3.2.17
django-environ==0.9.0
django-ckeditor==6.5.1
django-cleanup==7.0.0
django-js-asset==2.0.0
django_scheduler-0.10.1
pillow==9.4.0
pytz==2022.7.1
sorl-thumbnail==12.9.0
sqlparse==0.4.3
